+++ connectors/twitchconnector.gd
@@ -1,8 +1,8 @@
 extends "res://connectors/baseconnector.gd"
 
-var display_name = "Twitch"
-var tooltip = "Connector to connect to the Twitch API / IRC chat"
-var icon = preload("res://connectors/twitchconnector.png")
-var type = "twitch"
+var display_name : String = "Twitch"
+var tooltip : String = "Connector to connect to the Twitch API / IRC chat"
+var icon : StreamTexture = preload("res://connectors/twitchconnector.png")
+var type : String = "twitch"
 
-var scene = "res://connectors/twitchconnector_ui.tscn"
+var scene : String = "res://connectors/twitchconnector_ui.tscn"