+++ overlay.gd
@@ -4,15 +4,15 @@ var OpenVRConfig
 var OpenVROverlay
 var target_size = Vector2()
 
-var SettingsNode = "MainWindow/SettingsScreen/General settings/MarginContainer/VBoxContainer/HBoxContainer/"
+var SettingsNode : String = "MainWindow/SettingsScreen/General settings/MarginContainer/VBoxContainer/HBoxContainer/"
 
-func _ready():
+func _ready() -> void:
 	# Get configuration object
 	OpenVRConfig = preload("res://addons/godot-openvr/OpenVRConfig.gdns").new()
 	OpenVRConfig.set_application_type(2) # Set to OVERLAY MODE = 2, NORMAL MODE = 1
 	
 	# Find the OpenVR interface and initialise it
-	var arvr_interface = ARVRServer.find_interface("OpenVR")
+	var arvr_interface : ARVRInterface = ARVRServer.find_interface("OpenVR")
 
 	if arvr_interface and arvr_interface.initialize():
 		# Make sure we run at 90 FPS for this
@@ -22,7 +22,6 @@ func _ready():
 		# Set our resolution for the 2D scene
 		target_size = arvr_interface.get_render_targetsize()
 		$VRViewport.size = target_size
-		$VRViewport/VRBackground.rect_size = target_size
 		
 		SignalManager.emit_signal("render_targetsize", target_size)
 		
@@ -58,36 +57,30 @@ func _ready():
 		SignalManager.emit_signal("vr_init", "error")
 					
 	SignalManager.connect("settings_changed", self, "_on_settings_changed")
-	
-	$VRViewport/VRBackground.color = SettingsManager.get_value("user", "overlay/color", DefaultSettings.get_default_setting("overlay/color"))
-	$VRViewport/VRBackground.modulate.a = SettingsManager.get_value("user", "overlay/opacity", DefaultSettings.get_default_setting("overlay/opacity"))
-	
-func _process(delta):
+
+func _process(delta : float) -> void:
 	if Input.is_action_just_pressed("center_on_hmd"):
 		# Calling center_on_hmd will cause the ARVRServer to adjust all tracking data so the player is centered on the origin point looking forward
 		ARVRServer.center_on_hmd(true, false)
 
-func _on_settings_changed():
-	$VRViewport/VRBackground.color = SettingsManager.get_value("user", "overlay/color", DefaultSettings.get_default_setting("overlay/color"))
-	$VRViewport/VRBackground.modulate.a = SettingsManager.get_value("user", "overlay/opacity", DefaultSettings.get_default_setting("overlay/opacity"))
-	
+func _on_settings_changed() -> void:
 	if OpenVROverlay:
 		OpenVROverlay.set_overlay_width_in_meters(SettingsManager.get_value("user", "overlay/size", DefaultSettings.get_default_setting("overlay/size")))
 		attempt_tracking()
 
-func _on_trackers_changed(tracker_name, tracker_type, tracker_id):
+func _on_trackers_changed(tracker_name : String, tracker_type : String, tracker_id) -> void:
 	attempt_tracking()
 	
 func attempt_tracking():
-	var position_x = SettingsManager.get_value("user", "overlay/position_x", DefaultSettings.get_default_setting("overlay/position_x"))
-	var position_y = SettingsManager.get_value("user", "overlay/position_y", DefaultSettings.get_default_setting("overlay/position_y"))
-	var position_z = SettingsManager.get_value("user", "overlay/position_z", DefaultSettings.get_default_setting("overlay/position_y"))
+	var position_x : float = SettingsManager.get_value("user", "overlay/position_x", DefaultSettings.get_default_setting("overlay/position_x"))
+	var position_y : float = SettingsManager.get_value("user", "overlay/position_y", DefaultSettings.get_default_setting("overlay/position_y"))
+	var position_z : float = SettingsManager.get_value("user", "overlay/position_z", DefaultSettings.get_default_setting("overlay/position_y"))
 	
-	var rotation_x = SettingsManager.get_value("user", "overlay/rotation_x", DefaultSettings.get_default_setting("overlay/rotation_x"))
-	var rotation_y = SettingsManager.get_value("user", "overlay/rotation_y", DefaultSettings.get_default_setting("overlay/rotation_y"))
-	var rotation_z = SettingsManager.get_value("user", "overlay/rotation_z", DefaultSettings.get_default_setting("overlay/rotation_z"))
+	var rotation_x : float = SettingsManager.get_value("user", "overlay/rotation_x", DefaultSettings.get_default_setting("overlay/rotation_x"))
+	var rotation_y : float = SettingsManager.get_value("user", "overlay/rotation_y", DefaultSettings.get_default_setting("overlay/rotation_y"))
+	var rotation_z : float = SettingsManager.get_value("user", "overlay/rotation_z", DefaultSettings.get_default_setting("overlay/rotation_z"))
 	
-	var transform = Transform(Basis(Vector3(0, 0, 0)), Vector3(position_x, position_y, position_z))
+	var transform : Transform = Transform(Basis(Vector3(0, 0, 0)), Vector3(position_x, position_y, position_z))
 	transform.basis = transform.basis.rotated(Vector3(1, 0, 0), rotation_x)	
 	transform.basis = transform.basis.rotated(Vector3(0, 1, 0), rotation_y)	
 	transform.basis = transform.basis.rotated(Vector3(0, 0, 1), rotation_z)