+++ ui/dialogs/addconnectiondialog.gd
@@ -1,20 +1,20 @@
 extends WindowDialog
 
-var services = []
-var currentconnectorui = null
-var settings_dialog = null
+var services : Array = []
+var currentconnectorui : Node = null
+var settings_dialog : Node = null
 
-var ConnectorsBoxNode = "MarginContainer/VBoxContainer/ConnectorBoxNode"
-var ConnectorUIContainerNode = "MarginContainer/VBoxContainer/ConnectorUIContainer"
+var ConnectorsBoxNode : String = "MarginContainer/VBoxContainer/ConnectorBoxNode"
+var ConnectorUIContainerNode : String = "MarginContainer/VBoxContainer/ConnectorUIContainer"
 
 func _ready():
 	# Load dynamic connectors for services
-	var dir = Directory.new()
+	var dir : Directory = Directory.new()
 	
 	if dir.open("res://connectors") == OK:
 		dir.list_dir_begin()
-		var file_name = dir.get_next()
-		var blacklist = ["baseconnector.gd"]
+		var file_name : String = dir.get_next()
+		var blacklist : Array = ["baseconnector.gd"]
 
 		while (file_name != ""):
 			if dir.current_is_dir():
@@ -32,22 +32,22 @@ func _ready():
 	else:
 		print("Unable to load connectors - path not found!")	
 
-func edit_connection(connection):
+func edit_connection(connection : Dictionary):
 	for i in range(0, get_node(ConnectorsBoxNode).get_item_count()):
 		if get_node(ConnectorsBoxNode).get_item_metadata(i)["type"] == connection["type"]:
 			get_node(ConnectorsBoxNode).select(i)
 			_on_ConnectorBoxNode_item_selected(i)
 			currentconnectorui.set_connection_info(connection)
 
-func set_settings_dialog(dialog):
+func set_settings_dialog(dialog : Node):
 	settings_dialog = dialog
 
 func _on_OKButton_pressed():
 	currentconnectorui.verify_connection(self)
 
-func connection_verified(response):
+func connection_verified(response : Dictionary):
 	if response.error:
-		var accept_dialog = AcceptDialog.new()
+		var accept_dialog : AcceptDialog = AcceptDialog.new()
 		accept_dialog.window_title = "Authentication failed"
 		accept_dialog.dialog_text = "I am sorry - this token did not work! - " + response.message
 
@@ -60,8 +60,8 @@ func connection_verified(response):
 func _on_CancelButton_pressed():
 	visible = false	
 
-func _on_ConnectorBoxNode_item_selected(index):
-	var connector = get_node(ConnectorsBoxNode).get_item_metadata(index)
+func _on_ConnectorBoxNode_item_selected(index : int):
+	var connector : Dictionary = get_node(ConnectorsBoxNode).get_item_metadata(index)
 	
 	currentconnectorui = load(connector["scene"]).instance()
 	get_node(ConnectorUIContainerNode).add_child(currentconnectorui)