+++ ui/elements/options/baseoption.gd
@@ -1,4 +1,4 @@
 extends MarginContainer
 
-func set_label(text):
+func set_label(text : String) -> void:
 	$HBoxContainer/Label.text = text